Avoid starting app without Internet connection
Start NoInternetScreen when app is started with not Internet connection
On Internet connection lost, show NoInternetScreen, and go back to Home automatically when the connection is restored.
On permission denied open app settings and ask for permission when app is resumed.
Create a callbacks data class that will be passed to Screen Composable to avoid strong coupling between the ViewModel and compose function.